requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyverilog==1.3.0
joblib==1.3.1
pytest==7.4.0
jupyter==1.0.0
# Web application dependencies
Flask==2.3.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
# Image processing
Pillow==10.0.0
requests==2.31.0 